feat : 회원가입 부가요소 기본 입력 반영
