--- service-user/src/main/java/com/team13/serviceuser/controller/UserController.java
@@ -1,36 +1,0 @@
-package com.team13.serviceuser.controller;
-
-import com.team13.serviceuser.dto.JoinRequest;
-import com.team13.serviceuser.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/jwt-loginapi")
-public class UserController {
-    private final UserService userService;
-
-    @PostMapping("/join")
-    public String join(@RequestBody JoinRequest joinRequest) {
-
-        // loginId 중복 체크
-        if (userService.checkEmailDuplicate(joinRequest.getEmail())) {
-            return "로그인 아이디가 중복됩니다.";
-        }
-        // 닉네임 중복 체크
-        if (userService.checkNicknameDuplicate(joinRequest.getNickname())) {
-            return "닉네임이 중복됩니다.";
-        }
-        // password와 passwordCheck가 같은지 체크
-        if (!joinRequest.getPassword().equals(joinRequest.getPasswordCheck())) {
-            return "바밀번호가 일치하지 않습니다.";
-        }
-
-        userService.join(joinRequest);
-        return "회원가입 성공";
-    }
-}
