+++ service-user/src/main/java/com/team13/serviceuser/entity/User.java
@@ -20,21 +20,15 @@ public class User {
     @Column(name = "email", nullable = false, length = 50)
     private String email;
 
-    @Column(name = "password", nullable = false, length = 20)
+    @Column(name = "password", nullable = false, length = 60)
     private String password;
 
     @Column(name = "nickname", nullable = false, length = 20)
     private String nickname;
 
-//    @Column(name = "user_rating", nullable = false)
-//    private float userRating;
-//
-//    @Column(name = "location_bcode", nullable = false)
-//    private int locationBcode;
-//
-//    @Column(name = "location_address", nullable = false, length = 100)
-//    private String locationAddress;
-//
-//    @Column(name = "profile_image_url", nullable = false, length = 200)
-//    private String profileImageUrl;
+    @Column(name = "user_rating", nullable = false)
+    private float userRating;
+
+    @Column(name = "profile_image_url", nullable = false, length = 200)
+    private String profileImageUrl;
 }
