+++ service-user/src/main/resources/schema.sql
@@ -3,12 +3,10 @@ USE n1;
 CREATE TABLE IF NOT EXISTS users(
                        id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
                        email VARCHAR(50) NOT NULL,
-                       password VARCHAR(20) NOT NULL,
+                       password VARCHAR(60) NOT NULL,
                        nickname VARCHAR(20) NOT NULL,
---                        user_rating FLOAT NOT NULL,
---                        location_bcode INT NOT NULL,
---                        location_address VARCHAR(100) NOT NULL,
---                        profile_image_url VARCHAR(200) NOT NULL
+                       user_rating FLOAT NOT NULL,
+                       profile_image_url VARCHAR(200) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS reviews (
